Revert "🌱 Implement settings page and adjusting font size & family (#281)"
This reverts commit 11c15328d70e9db87d275b39fc77f42663ac5426.
