+++ src/components/settings/SettingsPage.tsx
@@ -2,34 +2,42 @@ import React from 'react';
 import GoBackButton from 'components/layout/GoBackButton';
 import { Header } from 'components/mdxOverrides';
 import copy from 'consts/copy';
-import useFontSize from 'utils/useFontSize';
-import useFontFamily from 'utils/useFontFamily';
 
-const cssClasses = {
-  fontSizeButton:
-    'flex items-center justify-center text-white rounded-full p-2 w-8 h-8',
-  enabledButton: 'bg-blue-500 hover:bg-blue-600 cursor-pointer',
-  disabledButton: 'bg-gray-400 cursor-default',
-  fontButton:
-    'text-white bg-blue-500 hover:bg-blue-600 cursor-pointer p-2 rounded-lg',
-  resetButton:
-    'bg-orange-500 text-white hover:bg-orange-600 cursor-pointer p-2 rounded-lg mb-20',
+const getItem = key => {
+  try {
+    return localStorage.getItem(key);
+  } catch {}
+};
+
+const setItem = (key, value) => {
+  try {
+    return localStorage.setItem(key, value);
+  } catch {}
 };
 
 export default function SettingsPage() {
-  const { fontSize, increaseFontSize, decreaseFontSize, resetFontSize } =
-    useFontSize();
-  const { fontFamily, setSerif, setSansSerif, resetFontFamily } =
-    useFontFamily();
+  const [fontSize, setFontSize] = React.useState(
+    () => getItem('fontSize') ?? 16,
+  );
 
-  const resetSettings = () => {
-    resetFontSize();
-    resetFontFamily();
+  React.useEffect(() => {
+    const root = document.getElementsByTagName('html')[0];
+    root.style.setProperty('font-size', `${fontSize}px`);
+  }, [fontSize]);
+
+  const buttonClass =
+    'flex items-center content-center text-white bg-green hover:bg-green-600 rounded-full p-2 w-6 h-6 cursor-pointer';
+  const handleFontDecrease = () => {
+    setFontSize(Number(fontSize) - 2);
+    setItem('fontSize', fontSize);
   };
-
-  const isMaxSize = fontSize >= 60;
-  const isMinSize = fontSize <= 10;
-
+  const handleFontIncrease = () => {
+    setFontSize(Number(fontSize) + 2);
+    setItem('fontSize', fontSize);
+  };
+  // TODO read it on page load
+  // TODO Format buttons
+  // TODO Reset settings
   return (
     <>
       <div className="relative whitespace-pre-line">
@@ -39,59 +47,17 @@ export default function SettingsPage() {
         <Header>{copy.settings.headline}</Header>
       </div>
       <div className="section">
-        <div className="flex flex-col items-center content-center space-y-5 mb-20">
-          <div className="flex items-center content-center bg-gray-50 bg-opacity-60 p-4 rounded-lg mb-2">
-            <div>{copy.preview}</div>
+        <div className="flex items-center content-center space-x-2">
+          <div className="flex items-center content-center bg-gray-50 bg-opacity-60 p-4 rounded-lg">
+            <div>Rozmiar tekstu: {fontSize}px</div>
           </div>
-          <div className="mb-2 font-semibold self-start">
-            {copy.fontSizeLabel}
-          </div>
-          <div className="flex space-x-4">
-            <button
-              className={`${cssClasses.fontSizeButton} ${
-                isMinSize ? cssClasses.disabledButton : cssClasses.enabledButton
-              }`}
-              onClick={decreaseFontSize}
-              disabled={isMinSize}
-            >
-              -
-            </button>
-            <span className="text-lg">{fontSize} px</span>
-            <button
-              className={`${cssClasses.fontSizeButton} ${
-                isMaxSize ? cssClasses.disabledButton : cssClasses.enabledButton
-              }`}
-              onClick={increaseFontSize}
-              disabled={isMaxSize}
-            >
-              +
-            </button>
-          </div>
-          <div className="mb-2 font-semibold self-start">{copy.fontLabel}</div>
-          <div className="flex space-x-4">
-            <button
-              className={`${cssClasses.fontButton} ${
-                fontFamily.includes('Lato') ? 'bg-blue-700' : ''
-              }`}
-              style={{ fontFamily: 'Lato, sans-serif' }}
-              onClick={setSansSerif}
-            >
-              Lato
-            </button>
-            <button
-              className={`${cssClasses.fontButton} ${
-                fontFamily.includes('NotoSerif') ? 'bg-blue-700' : ''
-              }`}
-              style={{ fontFamily: 'NotoSerif, serif' }}
-              onClick={setSerif}
-            >
-              NotoSerif
-            </button>
-          </div>
+          <button className={buttonClass} onClick={handleFontDecrease}>
+            -
+          </button>
+          <button className={buttonClass} onClick={handleFontIncrease}>
+            +
+          </button>
         </div>
-        <button className={cssClasses.resetButton} onClick={resetSettings}>
-          {copy.resetSettings}
-        </button>
         <GoBackButton />
       </div>
     </>
