+++ tailwind.config.js
@@ -75,20 +75,6 @@ module.exports = {
         800: '#2A2A2A',
         900: '#121212',
       },
-      blue: {
-        DEFAULT: '#3aa4d9',
-        50: '#f2f9fd',
-        100: '#e4f0fa',
-        200: '#c2e3f5',
-        300: '#8ccced',
-        400: '#4fb1e1',
-        500: '#3aa4d9',
-        600: '#197ab0',
-        700: '#16618e',
-        800: '#165276',
-        900: '#184562',
-        950: '#102c41',
-      },
       dark: '#000000',
       white: '#FFFFFF',
       transparent: 'transparent',
